Cloud fn to get authoration code and exchange it other tokens
The other tokens we get are the access and the refresh tokens.

We persist these tokens in Firebase. We will require these when we fetch 
events for the user after the initial sign up and whenever a new event 
is added.

We are using firebase cloud function environment variabkle to store the 
details of our Google Sign In auth client.