Functions related to calendar events webhooks
-  subcribe to a calendar(events list) by creating a new channel

-  handle new event notifiction
   -  get new events for the week
   -  re-run aggregates for the week

We need to register the domain on which the webhook will POST the 
notification.
The registration reuires a verification which is achieved by sending a 
HTML with a specific meta tag on a GET request to the webhook url.

Of course, we subscribe to the webhooks whenever a user signs up.